+++ src/FluxFlow.Components.Timers/Options/TimerComponentOptions.cs
@@ -1,0 +1,5 @@
+namespace FluxFlow.Components.Timers.Options;
+
+public sealed class TimerComponentOptions
+{
+}