+++ src/FluxFlow.Components.Timers/TimerComponentPorts.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.Timers;
+
+public static class TimerComponentPorts
+{
+    public const string Output = "Output";
+}