+++ src/FluxFlow.Components.Timers/TimerComponentTypes.cs
@@ -1,0 +1,8 @@
+using FluxFlow.Engine.Definitions;
+
+namespace FluxFlow.Components.Timers;
+
+public static class TimerComponentTypes
+{
+    public static readonly NodeType Interval = new("timer.interval");
+}