+++ src/FluxFlow.Components.Timers/TimerErrorCodes.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.Timers;
+
+public static class TimerErrorCodes
+{
+    public const int IntervalFailed = 9000;
+}